requirements.txt
accelerate==0.21.0
peft==0.4.0
bitsandbytes==0.39.0
loguru==0.7.0
numpy==1.21.4
pandas==1.2.5
tqdm==4.62.3
deepspeed==0.9.5
tensorboard
sentencepiece
transformers_stream_generator
tiktoken
einops
torch==2.0.0
